+++ Classes/Domain/Repository/MessagesRepository.php
@@ -22,8 +22,10 @@
 
 namespace Ubl\Supportchat\Domain\Repository;
 
+use TYPO3\CMS\Core\Utility\GeneralUtility;
 use TYPO3\CMS\Extbase\Persistence\QueryInterface;
 use TYPO3\CMS\Extbase\Persistence\Generic\QuerySettingsInterface;
+use TYPO3\CMS\Extbase\Persistence\Generic\Typo3QuerySettings;
 use TYPO3\CMS\Extbase\Persistence\Repository;
 
 
@@ -43,7 +45,8 @@ class MessagesRepository extends Repository
     public function initializeObject()
     {
         /** @var QuerySettingsInterface $querySettings */
-        $querySettings = $this->objectManager->get(QuerySettingsInterface::class);
+        $querySettings = GeneralUtility::makeInstance(Typo3QuerySettings::class);
+        //$querySettings = $this->objectManager->get(QuerySettingsInterface::class);
         $querySettings->setRespectStoragePage(false);
         $this->setDefaultQuerySettings($querySettings);
     }